[Refator] 폴더 재구조화
