+++ public/utils/utils.js
@@ -1,0 +1,9 @@
+const delay = (delayTime) => {
+  return new Promise((resolve, reject) => {
+    setTimeout(() => {
+      resolve("Success");
+    }, delayTime);
+  });
+};
+
+export default delay;
